requirements.txt
asgiref>=3.3.0
celery==5.0.5
cssselect==1.1.0
Django==3.2.16
django-elasticsearch-dsl==7.2.2
django-filter==2.4.0
djangorestframework==3.12.2
factory-boy==3.2.0
pikepdf==5.6.1
playwright==1.27
polib==1.1.0
psycopg2==2.8.6
python-dotenv==0.15.0
redis==3.5.3
requests==2.25.1
sentry-sdk==0.19.5
uWSGI==2.0.19.1
